pyproject.toml
[tool.poetry]
name = "itfox-tpb"
version = "0.1.0"
description = ""
authors = ["NovaCript <glebka.yatsenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
